+++ bufflogin/buff.py
@@ -1,7 +1,7 @@
 from typing import (
     Any,
     Dict,
-    Type,
+    Optional,
 )
 
 import lxml.html as html
@@ -20,24 +20,18 @@ from .schemas import BuffAuthorizationStatus
 
 
 class Buff:
-    """
-    Authorization to buff.163.com through Steam.
-    """
     def __init__(
         self,
         steam: Steam,
-        cookie_storage: Type[CookieStorageAbstract] = BaseCookieStorage,
-        request_strategy: Type[RequestStrategyAbstract] = BaseRequestStrategy,
+        cookie_storage: Optional[CookieStorageAbstract] = None,
+        request_strategy: Optional[RequestStrategyAbstract] = None,
     ):
         self._steam = steam
-        self._http = request_strategy()
-        self._storage = cookie_storage()
+        self._requests = request_strategy if request_strategy is not None else BaseRequestStrategy()
+        self._storage = cookie_storage if cookie_storage is not None else BaseCookieStorage()
 
     async def request(self, url: str, method: str = 'GET', **kwargs: Any) -> str:
-        """
-        Request with Buff session.
-        """
-        return await self._http.text(
+        return await self._requests.text(
             url=url,
             method=method,
             cookies=await self._storage.get(self._steam.login, domain='buff.163.com'),
@@ -45,18 +39,12 @@ class Buff:
         )
 
     async def is_authorized(self) -> bool:
-        """
-        Is alive authorization.
-        """
         response: str = await self.request(
             url='https://buff.163.com/api/market/goods/bill_order',
         )
         return BuffAuthorizationStatus.parse_raw(response).is_authorized()
 
     def parse_openid_params(self, response: str) -> Dict[str, str]:
-        """
-        Parse openid parameters.
-        """
         page = html.document_fromstring(response)
         params = {
             'action': '',
@@ -69,10 +57,7 @@ class Buff:
         return params
 
     async def get_openid_params(self) -> Dict[str, str]:
-        """
-        Get openid parameters for authorization.
-        """
-        response = await self._http.text(
+        response = await self._requests.text(
             method='GET',
             url='https://buff.163.com/account/login/steam?back_url=/',
             cookies=await self._steam.cookies(),
@@ -80,13 +65,10 @@ class Buff:
         return self.parse_openid_params(response)
 
     async def login_to_buff(self) -> None:
-        """
-        Login to buff.163.com.
-        """
         if await self.is_authorized():
             return
         await self._steam.login_to_steam()
-        await self._http.bytes(
+        await self._requests.bytes(
             method='POST',
             url='https://steamcommunity.com/openid/login',
             data=await self.get_openid_params(),
@@ -95,7 +77,7 @@ class Buff:
         await self._storage.set(
             login=self._steam.login,
             cookies={
-                'buff.163.com': self._http.cookies('buff.163.com'),
+                'buff.163.com': self._requests.cookies('buff.163.com'),
             },
         )
         if not await self.is_authorized():