+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# Bufflogin changelog
+# Changelog
+
+## [v0.0.4 (2023-01-22)](https://github.com/sometastycake/bufflogin/releases/tag/v0.0.4)
+
+- Removed docstrings
+- Buff class now accepts objects of `BaseRequestStrategy` and `BaseCookieStorage`
+- Upgraded `pysteamauth` library to `v1.0.0` version
 
 ## [v0.0.3 (2022-11-15)](https://github.com/sometastycake/bufflogin/releases/tag/v0.0.3)
 
@@ -10,4 +16,4 @@
 
 ## [v0.0.1 (2022-09-15)](https://github.com/sometastycake/bufflogin/releases/tag/v0.0.1)
 
-- Authorization to buff.163.com through Steam
+- Authorization to `buff.163.com` through Steam